LPS-186836 Extract service from AnalyticsConfigurationRegistryImpl to AnalyticsConfigurationManagedServiceFactory, create AnalyticsConfigurationRegistryUtil and apply to usages
